Native (`date`, iOS): Implement gap handling
Sometimes an operation on dates leads to an invalid datetime: when
clock is moved forward (to account for DST, for example), an hour
is skipped, an all the datetimes inside that hour are considered
invalid. When this happens, it is considered by most sources that
such a date is expressed in terms of the offset before the
transition, and to fix it, one needs to add to it the duration of
the transition. This is now implemented.