Navite (all): additionally improve performance.
The tactics employed in this commit are questionable.

The main change is that now, instead of passing strings between
Kotlin and native parts, numeric ids are used. To have a
sensible interpretation of such ids that prohibited their
invalidation, Apple and Windows use additional memory to store the
mapping. Moreover, it is assumed for Apple that the set of
timezones and their rules never change, and for *nix, it is
assumed that the timezone objects are always at the same memory
location, which is true for now but will change if the
implementation switches to use C++20 capabilities.