Native (all): fix (unneeded) range checks on internal constructors
There are some constructors that were only invoked internally and
in such ways that the argument requirements could not be violated.
However, just in case we ever accidentally use these constructors
in a non-safe way, it's probably a good idea to have these checks
work correctly, given that they are not expensive.