Native (Kotlin): add missing range checks (#11)
`LocalDateTime` could be constructed with invalid time values,
which is problematic. Additionally, `Instant` and `LocalDate` didn't
throw exceptions when their values were valid but outside of the
supported range.

Also, there are some constructors that were only invoked internally
in such ways that the argument requirements could not be violated.
However, just in case we ever accidentally use these constructors
in a non-safe way, it's probably a good idea to have these checks
work correctly, given that they are not expensive.