Native: run tests in a test worker
Leaving out `@SharedImmutable` or `@ThreadLocal` on field
declarations causes Kotlin/Native implementation to fail when
the program is run in a multithreaded environment. In order to
detect these omissions, we now run all tests in a test worker
(that is, in a separate thread), which causes the program to crash
if a field without a thread safety marker.