Common: TimeZone.SYSTEM -> TimeZone.currentSystemDefault()
In addition to the renaming, caching of the system timezone was
disabled. Before, due to querying the timezone being an expensive
operation (especially on Native), we acquired the system timezone
only once. The main motivation behind it was that having an
expensive operation on a `val` was against the expectations of
users, and most JRE implementations do not update the system
timezone.

However, it turned out that at least on Android the changes to
system timezone are tracked by the Java runtime, so it is not
viable to just assume that the system timezone won't change. Thus,
we decided to perform the renaming and thus communitate that the
operation may be an expensive one.