Native (Darwin): improve results outside of supported range
Darwin only supports years [1; 4001], and strange things may happen
when requesting from it to perform arithmetic on dates outside this
range. In particular, when converting NSDateComponents to NSDate,
the era is ignored, and so 32000 BC becomes 32000 AD, which is
absolutely wrong.

Reliance on Darwin auto-adjusting NSDate when converting from
NSDateComponents is the core of our solution to acquire the
timezone database information, but in this case it is better to
completely ignore this API and instead return a plausible-looking
result.