Native (Windows): return more plausible results
When querying the timezone database outside of the supported range
on Windows, the requested value could be too large to fit in the
platform-provided range. Then, instead of returning a completely
arbitrary result, the call will still return values that could be
naturally acquired for some date in the supported range, instead
of garbage values.