Native: make range for Instants and LocalDate a 1000 times narrower (#22)
This is done to support `LocalDate.daysUntil(...): Int` always
having a precise value. Additionally, it allows to simplify some
parts of the implementation.

Before, the Native implementation was using the same representable
ranges as the Java implementation; now, it uses the same ranges as
the JS implementation, which are narrower so that everything fits
in a JS number.