Fix adding small Duration to large Instant on JS and Native
When adding a positive Duration < 1 second to an Instant far in the
future for which the addition implementation overflows and throws an
exception, the catch blocks would only check if the seconds component of
the Duration is > 0. However since the Duration is < 1 second, the
seconds component is not > 0 and Instant.MIN is returned instead of
Instant.MAX.

This was fixed by replacing the check against the seconds component of
the Duration with a check against the whole Duration like it was already
done in the JVM implementation.

See #263