Ensure better ISO 8601 compliance
ISO 8601 defines the period of one month as a fixed number of
calendar days, specified by the calendar.

Before this change, we treated a month as a separate time scale
unit during `DatePeriod` arithmetics, namely, we resolved the
`LocalDateTime` on which we did arithmetics internally after adding
the months but before adding the days.

This change makes it so that adding months and days is done as a
single change, only consulting the calendar, and the time zone is
only used once all calendar-based operations are performed.

The code that calculates the `DateTimePeriod` between two
`Instant` values is changed symmetrically.