Use a single `parse` overload with a default parameter
Before this change, we had two `parse` overloads: one that accepted
an ISO string and one that accepted some string and the
corresponding formatter. Now, we only have one overload that uses
the ISO format by default.

As a workaround for KT-65484, we have to hide the actual default
parameters behind a redundant `internal` function.