Bump Gradle version to 8.9
Rationale is a slightly better DX when working with a version catalog (which is to be introduced in a later commit).

Usages of Property API now leverage the assignment Kotlin compiler plugin instead of invoking `set` methods directly.

Usages of deprecated Gradle API are either replaced with supported alternatives or removed due to not really doing anything.