Experiment with #178
... And fail. If the library uses `js-joda/@timezone`
as a compile-time dependency, then any projects depending on the
library will fail to build if they don't have the same
dependency, as this commit shows: try
`./gradlew publishToMavenLocal` and then, in the
`js-without-timezones` directory, `./gradlew check` will fail.

Another option would be to use `require()` from the library.
This would allow us to access the `js-joda/@timezone` dependency
without declaring that we depend on it:
```
js("{ try { require('@js-joda/timezone'); return null; } catch(e) { return e; } }")
```
However, the Kotlin/JS team discouraged this and explained to me
that the `require` approach was outdated in the JS ecosystem.