Introduce LocalIsoWeekDate
This is a new entity with the narrow use case of representing
ISO week dates.
It is not suitable for datetime arithmetics, and its week number
behaves extremely unintuitively around the edges of a normal
calendar year.
It is not advised to use this class for anything but its exact
stated purpose: as transport format for `LocalDate` in scenarios
that require the specific ISO week-date format.